This function generates parameters/settings for ensemble deconvolution

This function generates parameters/settings for ensemble deconvolution.

get_params(
  TNormalization = c("CPM", "none", "TPM", "TMM", "QN"),
  CNormalization = c("CPM", "none", "TPM", "TMM", "QN"),
  Scale = c("log", "linear"),
  data_type,
  data_name,
  n_markers = 50,
  Marker.Method = c("t", "wilcox", "combined", "none", "p.value", "regression"),
  dmeths = NULL,
  teqc = TRUE,
  batchcorrec = FALSE,
  RB_only = TRUE
)

Arguments

Scale

Scaling methods for gene expression data.

(Option) Vector of string. Defaults are "log" and "linear"

data_type

Reference data data type

(Required) One-dimensional string. Options are

'singlecell-rna' Single cell RNA seq reference.

'microarray' Microarray reference.

data_name

Name for bulk_reference

(Options) One-dimensional string. Can be in format : bulk_reference

n_markers

Number of markers

(Option) One-dimensional numeric, default is 150

How many markers genes to use for deconvolution. Can either be a single integer, vector of integers. All cell types use the same number of markers.

dmeths

Deconvolution methods to used. Default is NULL, EnsDeconv will run all possible deconvolution methods for specific reference data sets.

Marker.Method

Method used to choose marker genes

'ratio' selects and ranks markers by the ratio of the mean expression of each gene in each cell type to the mean of that gene in all other cell types.

'regression ' selects and ranks markers by estimated regression coefficients in a series of regressions with single covariate that is indicator of each type.

'diff' selects and ranks markers based upon the difference, for each cell type, between the median expression of a gene by each cell type and the median expression of that gene by the second most highly expressed cell type.

'p.value' selects and ranks markers based upon the p-value of a t-test between the median expression of a gene by each cell type and the median expression of that gene by the second most highly expressed cell type.

't' Perform pairwise Welch t-tests between groups of cells, possibly after blocking on uninteresting factors of variation.

'wilcox' Perform pairwise Wilcoxon rank sum tests between groups of cells, possibly after blocking on uninteresting factors of variation.

'binom' Perform pairwise binomial tests between groups of cells, possibly after blocking on uninteresting factors of variation.

'TOAST' Iteratively searches for cell type-specific features

'Hedge' Iteratively searches for cell type-specific features

'roc' Iteratively searches for cell type-specific features

teqc

Logical. Use same normalization between bulk data and reference data. Default: TRUE.

batchcorrec

Logical. Apply batch correction or not. Default: FALSE

Normalization

normalization methods for bulk gene expression and reference gene expression

(Option) Vector of string. Defaults are "CPM", "TPM" and "none"

Value

a data.frame that each row corresponds to a specific deconvolution scenario/setting