This function is used for process output

This function is used for process output

process_all(
  allfold_res,
  allgene_res,
  df_true = NULL,
  trueMet = NULL,
  feature_selection = TRUE,
  lambda = NULL,
  incl_lasso = FALSE,
  get_metric = TRUE,
  rm_dmeths = TRUE,
  r2 = FALSE,
  r2_ind = NULL,
  scale = "log",
  transformation = "CPM",
  optimization = FALSE
)

Arguments

allfold_res

List of output from gen_all_res or gen_all_res_list of two fold

Each element of list is result for one scenario, which contains element a, p, and ensemble.

'a' Contains p_hat (list, deconvolution results of each fold), p_truth (list, true cell type proportion), n (list, number of markers of each fold), time, markers (list, markers gene of each fold), all_time (computation time)

'p' Parameters

'ensemble' Testing data for each fold

allgene_res

List of output from gen_all_res or gen_all_res_list of one fold

Each element of list is result for one scenario, which contains element a, p, and ensemble.

'a' Contains p_hat (list of deconvolution results), p_truth (list, true cell type proportion), n (list, number of markers of each fold), time, markers (list, markers gene of each fold), all_time (computation time)

'p' Parameters

'ensemble' Testing data for

df_true

Dataframe, True cell type expression, samples*cell type

trueMet

Character, name of bulk data

feature_selection

Logical, conducting feature selection or not

lambda

(Optional) tuning parameter for feature selection

get_metric

Logical, calculating metric or not

rm_dmeths

Logical, remove several deconvolution methods or not when ensemble